Added main blend file wih python and Sunfish logic
This is the first upload of the BGE based chess game.  It is best to run Blender with the console available to show CLI information.  With this commit, moves aren't being displayed on the chess board.  This build only showcases integration of Sunfish.py and how BGE connectivity can be made.